Split up VectorBase to seperate memory management out from it
